% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mapshaper.R
\name{renderMapshaper}
\alias{renderMapshaper}
\title{Widget render function for use in Shiny}
\usage{
renderMapshaper(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

