% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mapshaper.R
\name{mapshaper}
\alias{mapshaper}
\title{Mapshaper in R}
\usage{
mapshaper(..., width = "100\%", height = "100\%")
}
\arguments{
\item{...}{One or many input lists, data.frame, or spatial class.  If \code{input}
  is not \code{json}, then automatic conversion will be attempted
  with \code{\link[geojsonio]{geojson_json}}.}
}
\description{
Use \href{www.mapshaper.org}{mapshaper} within R to analyze and simplify
your spatial objects.
}
\examples{
library('geojsonio')
library('sp')
library('mapshaperWidget')

###  example from geojsonio vignette
poly1 <- Polygons(
  list(
    Polygon(cbind(c(-100,-90,-85,-100), c(40,50,45,40)))
  )
  , "1"
)
poly2 <- Polygons(
  list(
    Polygon(cbind(c(-90,-80,-75,-90),c(30,40,35,30)))
  )
  , "2"
)
sp_poly <- SpatialPolygons(list(poly1, poly2), 1:2)

mapshaper( sp_poly )


### how to provide multiple spatial to mapshaper
calif <- geojson_read(
  system.file("examples", "california.geojson", package = "geojsonio")
)
mapshaper( sp_poly, calif )
}

